Refactor click module names (for myclick and utils). Add docstrings.
