+++ nameko/click_cli/__init__.py
@@ -1,67 +1,16 @@
-from functools import partial
-
 import click
 
 from nameko import config
 from nameko.exceptions import CommandError, ConfigurationError
 
-from .utils.config import setup_config
-from .utils.paramtypes import HOST_PORT, KEY_VAL, NamekoModuleServicesParamType
-
-
-def flatten_list_of_lists(ctx, param, list_of_lists):
-    return sum(list_of_lists, [])
-
-
-# CLI decorators. Allow reuse of unified opts/args whenever needed.
-option_broker = partial(
-    click.option,
-    "--broker",
-    help="Deprecated option for setting up RabbitMQ broker URI.\n"
-    "Use --define or --config and set AMQP_URI instead.",
-    metavar="BROKER",
-)
-option_config_file = partial(
-    click.option,
-    "-c",
-    "--config",
-    "config_file",
-    help="The YAML configuration file",
-    type=click.File("rb"),
-    metavar="CONFIG",
-)
-option_define = partial(
-    click.option,
-    "-d",
-    "--define",
-    help="Set config entry. Overrides value loaded from config file."
-    " Can be used multiple times. Example: --define"
-    " AMQP_URI=pyamqp://guest:guest@localhost",
-    type=KEY_VAL,
-    multiple=True,
-    metavar="KEY=VALUE",
-    callback=lambda ctx, param, value: dict(value)
-)
-option_backdoor_port = partial(
-    click.option,
-    "--backdoor-port",
-    help="Specify a port number to host a backdoor, which can be connected to"
-    " for an interactive interpreter within"
-    " the running service process using `nameko backdoor`.",
-    type=HOST_PORT,
-)
-option_interface = partial(
-    click.option, "--interface", type=click.Choice(["bpython", "ipython", "plain"])
-)
-argument_services = partial(
-    click.argument,
-    "services",
-    nargs=-1,
-    required=True,
-    metavar="module[:service class]",
-    # allow import of modules from current dir (manipulate sys.path)
-    type=NamekoModuleServicesParamType(extra_sys_paths=["."]),
-    callback=flatten_list_of_lists,
+from .utils import setup_config
+from .myclick.argsopts import (
+    option_broker,
+    option_config_file,
+    option_define,
+    option_backdoor_port,
+    option_interface,
+    argument_services,
 )
 
 
@@ -108,6 +57,7 @@ positional arguments:
     """
 
     from .run import main
+
     try:
         setup_config(config_file, define, broker)
         if backdoor_port:
