+++ nameko/click_cli/utils/__init__.py
@@ -1,0 +1,5 @@
+"""Importing all relevant utility functions.
+"""
+from .config import setup_config  # noqa
+from .code import interact  # noqa
+from .import_services import import_services  # noqa
